Don't delete united series till no readers remain

+++ store/src/main/scala/ionic/store/series/SeriesParceler.scala
@@ -1,5 +1,7 @@
 package ionic.store.series
 
+import com.google.common.collect.HashMultiset
+import com.google.common.collect.Multiset
 import java.util.concurrent.Executor
 import java.util.concurrent.Executors
 
@@ -17,26 +19,17 @@ import org.apache.avro.generic.GenericRecord
 import com.threerings.fisy.Directory
 import com.threerings.fisy.impl.local.LocalDirectory
 
-class ParceledReader(query: Query, root: Directory, writers: Iterable[UnitedSeriesWriter],
-  splits: Iterable[Directory]) extends Iterable[GenericRecord] {
-  val name = query.from
-  // TODO - thread safety of written
-  val openPositions: Iterable[Tuple2[Directory, Long]] =
-    writers.map((w: UnitedSeriesWriter) => { (w.dest, w.written) })
-  def iterator(): Iterator[GenericRecord] = {
-    (splits.flatMap(new SplitSeriesReader(_, query.where)) ++
-      openPositions.flatMap((t: Tuple2[Directory, Long]) => {
-        new UnitedSeriesReader(t._1, query.where, t._2)
-      })).iterator
-  }
-}
 class SeriesParceler(val base: LocalDirectory, name: String) extends Logging {
   import ionic.util.ReactImplicits._
 
-  private val writers = new HashSet[UnitedSeriesWriter]
   private val splitter :Executor = Executors.newSingleThreadExecutor()
-  private def splits: Iterable[Directory] =
-    base.navigate(Series.splitPrefix + "/" + name).collect({ case d: Directory => d })
+
+  // open series writers, readers on those writers, and fully-transferred splits. All access to
+  // these variables post-construction must be inside synchronization on writers.
+  private val writers = new HashSet[UnitedSeriesWriter]
+  private val openUnited :Multiset[Directory] = HashMultiset.create()
+  private val splits: Buffer[Directory] =
+    base.navigate(Series.splitPrefix + "/" + name).collect({ case d: Directory => d }).toBuffer
 
   splits.map(new SplitSeriesReader(_)).filter((reader: SplitSeriesReader) => {
     base.navigate(reader.meta.transferredFrom.toString).exists()
@@ -49,10 +42,11 @@ class SeriesParceler(val base: LocalDirectory, name: String) extends Logging {
       unitedDir.delete()
     } else {
       log.warn("Found incomplete transfer to split. Deleting", "split", split.source)
+      splits -= split.source
       split.source.delete()
       if (united.entries > 0) {
         log.warn("Redoing incomplete transfer")
-        SplitSeriesWriter.transferFrom(base, united)
+        splits += SplitSeriesWriter.transferFrom(base, united).dest
       } else {
         log.warn("Incomplete transfer was of empty united, deleting it as well",
           "united", unitedDir)
@@ -62,12 +56,47 @@ class SeriesParceler(val base: LocalDirectory, name: String) extends Logging {
   })
   base.navigate(Series.unitedPrefix + "/" + name).collect({ case d: Directory => d }).map(
     new UnitedSeriesReader(_)).foreach((reader :UnitedSeriesReader) => {
-      SplitSeriesWriter.transferFrom(base, reader)
+      splits += SplitSeriesWriter.transferFrom(base, reader).dest
       reader.source.delete()
     })
 
-  def reader(clauses: String = ""): Iterable[GenericRecord] =
-    new ParceledReader(Query.parse(name), base, writers, splits)
+  private def readers(query :Query) :Iterable[Iterator[GenericRecord]] = {
+    writers synchronized {
+      val openPositions: Iterable[Tuple2[Directory, Long]] =
+        writers.map((w: UnitedSeriesWriter) => { (w.dest, w.written) })
+      (splits.map(new SplitSeriesReader(_, query.where)) ++
+        openPositions.map((t: Tuple2[Directory, Long]) => {
+          openUnited.add(t._1)
+          new UnitedSeriesReader(t._1, query.where, t._2)
+        }))
+    }
+  }
+  private def release(iterators :Iterable[Iterator[GenericRecord]]) {
+    writers synchronized {
+      iterators.collect({case i: UnitedSeriesReader => i}).foreach(
+        (reader :UnitedSeriesReader) => {
+          openUnited.remove(reader.source)
+          if (!openUnited.contains(reader.source)) reader.source.delete()
+          })
+      }
+  }
+  def reader(clauses: String = "") = new CloseableIterable[GenericRecord] {
+    val query = Query.parse(name)
+    def iterator() = new CloseableIterator[GenericRecord] {
+      var closed = false
+      val iterators = readers(query)
+      val flattened = iterators.iterator.flatten
+      def hasNext :Boolean = if (!flattened.hasNext) {
+        close()
+        return false
+      } else return true
+      def next =  flattened.next
+      def close = if (!closed) {
+          closed = true
+          release(iterators)
+        }
+    }
+  }
 
   def writer(schema: Schema): UnitedSeriesWriter = {
     val writer = new UnitedSeriesWriter(schema, base)
@@ -75,14 +104,15 @@ class SeriesParceler(val base: LocalDirectory, name: String) extends Logging {
       // TODO - thread safety of writer transfer
       splitter.execute(new Runnable() {
         def run {
-          // TODO - don't make visible until transfer completed
-          SplitSeriesWriter.transferFrom(base, writer)
-          writers -= writer
-          // TODO - don't delete if there's an open reader
-          writer.dest.delete()
+          val split = SplitSeriesWriter.transferFrom(base, writer)
+          writers synchronized {
+            splits += split.dest
+            writers -= writer
+            if (!openUnited.contains(writer.dest)) writer.dest.delete()
+          }
         }})
     })
-    writers += writer
+    writers synchronized { writers += writer }
     writer
   }
 }
